Add files via upload
This is a program that I wrote to learn how to web scrape. I used common resources online to help me develop this program. I was inspired by github user lazargugleta and from there used a slightly different combination of libraries to gather information regarding the coronavirus update. Feel free to use this code yourself and improve on it.